Formulize and contain entire danggit build.
We'll use this in CI, and forgo all attempts to have any faith state of the CI host system.

As the comments in the formula file say, it would be even better to directly address the problematic libraries... but it also doesn't hurt a bit to have this in place for the GCC issue, which is hard to tackle with any smaller of a sledgehammer anyway, so here we go.

This doesn't work because the container doesn't include git :D  But we shouldn't have an unconditionnal dependency on that for testing anyway, so I'll fix that shortly.

Signed-off-by: Eric Myhre <hash@exultant.us>