Attempting simple commit func.
Taking the entire filesystem tree as a param of in-memory blobs.  This worked effectively in repeatr for test fixtures, so copying that pattern here.  Not necessarily suitable for performant operations on real repos and large amounts of data; not meant to be.

Probably doesn't work correctly on actual trees with dirs yet.  Committing as WIP.  (Need to go Do Something about more clib stuff because building this on my host is a PITA.)

Signed-off-by: Eric Myhre <hash@exultant.us>