Try committing dirs.  Incomplete.
Dealing with actual recursive trees is... fairly irritating; you have to sort it all out and then do a treewalk with tree and post visits and keep stacks.  This stuff is written over in repeatr's tar hash impl, not coincidentally.

But I just don't care enough to drag that over here right now, given that this is for rough testing stuff and not yet a feature.  Bummer that the libraries don't have a reusable piece of this.

Signed-off-by: Eric Myhre <hash@exultant.us>