Refnames deserve a type.
Just putting it in place for now so it's available for API clarity, but it may also sprout helper methods in the future (e.g. to translate that "refs/heads/master" is usually just printed "master").

(Ths change is coming in as result of work in parallel in another project, where I'm using this API explicitly to exercise it, though the project has no libgit concerns at all (yet).)

Signed-off-by: Eric Myhre <hash@exultant.us>